Move managers to “util” package
